Add GDT and IDT initialization code
On a call to cpu.init(), the cpu will initialize and load the run-time
GDT and IDT. This causes a stack switch to occur on any of the ISA
exceptions (any interrupt < 32). Currently all interrupt handlers dump
all registers and abort.

For cpu 0, the initialization is done early, the interrupt stack is
statically allocated and therefore the cpu init can be done very early
to catch any exceptions or faults early on.

Change-Id: Ic3010b38953dfd8a75cc99906f23e397124bb8bb