Make assert macro "use" variable if turned off
When NDEBUG is defined the assert macro doesn't actually do a
check. This can cause an unused variable warning if the value isn't used
anywhere but an assert statement. This commit uses a bit of a hack to
prevent the compiler from issuing the warning without actually
evaluating the expression inside the assert.

Change-Id: I63a20d7772ec7edd9274f5df7eea032691c83bca